Resolve "Metainfo support for units with scalar numeric quantities"

Closes #1006

See merge request nomad-lab/nomad-FAIR!811

+++ nomad/datamodel/metainfo/simulation/calculation.py
@@ -998,7 +998,7 @@ class DosFingerprint(MSection):
         ''')
 
     indices = Quantity(
-        type=np.dtype(np.int16),
+        type=np.dtype(np.int32),
         shape=[2],
         description='''
         Indices used to compare DOS fingerprints of different energy ranges.
